utils: added a custom "kmsprint"
Added a custom 'kmsprint' variant, which outputs the modes for the detected cards/connectors.
It behaves just like 'kmsprint -m', but with a different formatting for the video modes detected.
The only parameter kept from the original 'kmsprint' is the device path, for situations where a diferent card needs to be inquired.